+++ src/main/java/ch/ethz/idsc/owl/glc/std/StandardTrajectoryPlanner.java
@@ -6,7 +6,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Objects;
-import java.util.Optional;
 
 import ch.ethz.idsc.owl.data.GlobalAssert;
 import ch.ethz.idsc.owl.glc.core.AbstractTrajectoryPlanner;
@@ -107,13 +106,6 @@ public class StandardTrajectoryPlanner extends AbstractTrajectoryPlanner {
         domainQueue.remove();
       }
     }
-  }
-
-  @Override
-  protected Optional<GlcNode> getFurthestGoalNode() {
-    // TODO JONAS operation is not defined for standard planner, but some demos (deltaxt...) use it !!!
-    // throw new UnsupportedOperationException();
-    return Optional.empty();
   }
 
   @Override
